requirements.txt
django==3.1.7
psycopg2==2.8.6
python-dateutil==2.8.1